added components and views - setting up for login view
